+++ src/Game/RegionArray.php
@@ -17,4 +17,22 @@ class RegionArray extends LoadedArray
 	{
 		return $this->getOffsets();
 	}
+
+	/**
+	 * Returns alle regions owned by the specified owner(s).
+	 *
+	 * @param integer $owner one or multiple owner (see {@see RegionState} constants) logically combined,
+	 *                       example: `getRegions(RegionState::OWNER_ME | RegionState::OWNER_NEUTRAL)`
+	 *
+	 * @return RegionArray
+	 *
+	 */
+	public function filterOwner($owner)
+	{
+		return $this->filter(function ($_region) use ($owner)
+		{
+			/** @var Region $_region */
+			return $_region->getOwner() & $owner;
+		});
+	}
 }