+++ src/Game/SuperRegionArray.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace Prokki\Warlight2BotTemplate\Game;
+
+use Prokki\TheaigamesBotEngine\Util\ArrayObject\Filterable;
+use Prokki\TheaigamesBotEngine\Util\ArrayObject\LoadedArray;
+use Prokki\TheaigamesBotEngine\Util\ArrayObject\GetOffsetable;
+
+class SuperRegionArray extends LoadedArray
+{
+	use Filterable, GetOffsetable;
+
+	/**
+	 * @return integer[]
+	 */
+	public function getIds()
+	{
+		return $this->getOffsets();
+	}
+}